Rename and move utils
